requirements.txt
langchain-community
langchain
chromadb
ollama
flask
shiny
shinywidgets
pymupdf
pillow
tesseract
python-docx